Refactor functions that generate the fake requests to a file called testRequests.js
